Mag raw data can be up to 32 bits

+++ include/sensors/mag_sensor.h
@@ -51,11 +51,11 @@ typedef struct __MagSensor_Raw_Data {
     uint64_t Timestamp; 		//!< Time stamp count in usec
     uint16_t Sensitivity[3];	//!< Scale factor in nanoTesla of the sensor
     union {
-        int16_t Val[3];
+        int32_t Val[3];
         struct {
-            int16_t X;          //!< X axis
-            int16_t Y;          //!< Y axis
-            int16_t Z;          //!< Z axis
+            int32_t X;          //!< X axis
+            int32_t Y;          //!< Y axis
+            int32_t Z;          //!< Z axis
         };
     };
 } MagSensorRawData_t;
